Refactor PatchSet to be contained in a separate package for reusability.
