+++ cluster-autoscaler/simulator/common/patchset_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package snapshot
+package common
 
 import (
 	"maps"
@@ -22,308 +22,6 @@ import (
 
 	"k8s.io/utils/ptr"
 )
-
-func TestPatchOperations(t *testing.T) {
-	p := newPatch[string, int]()
-
-	// 1. Get and IsDeleted on non-existent key
-	if _, found := p.Get("a"); found {
-		t.Errorf("Get for 'a' should not find anything")
-	}
-	if p.IsDeleted("a") {
-		t.Errorf("IsDeleted for 'a' should be false")
-	}
-
-	// 2. Set 'a' key
-	p.Set("a", 1)
-	if val, found := p.Get("a"); !found || val != 1 {
-		t.Errorf("Get('a') = %v,%v, expected 1,true", val, found)
-	}
-	if p.IsDeleted("a") {
-		t.Errorf("IsDeleted('a') = true, should be false")
-	}
-
-	// 3. Overwrite 'a' key
-	p.Set("a", 2)
-	if val, found := p.Get("a"); !found || val != 2 {
-		t.Errorf("Get('a') = %v,%v, expected 2,true", val, found)
-	}
-	if p.IsDeleted("a") {
-		t.Errorf("IsDeleted('a') = true, should be false")
-	}
-
-	// 4. Delete 'a' key
-	p.Delete("a")
-	if val, found := p.Get("a"); found {
-		t.Errorf("Get('a') = %v,%v, should not find anything after delete", val, found)
-	}
-	if !p.IsDeleted("a") {
-		t.Errorf("IsDeleted('a') = false, should be true")
-	}
-
-	// 5. Set 'a' key again after deletion
-	p.Set("a", 3)
-	if val, found := p.Get("a"); !found || val != 3 {
-		t.Errorf("Get('a') = %v,%v, expected 3,true", val, found)
-	}
-	if p.IsDeleted("a") {
-		t.Errorf("IsDeleted('a') = true, should be false")
-	}
-
-	// 6. Delete a non-existent key 'c'
-	p.Delete("c")
-	if val, found := p.Get("c"); found {
-		t.Errorf("Get('c') = %v, %v, should not find anything", val, found)
-	}
-	if !p.IsDeleted("c") {
-		t.Errorf("IsDeleted('c') = false, should be true")
-	}
-}
-
-func TestCreatePatch(t *testing.T) {
-	tests := map[string]struct {
-		sourceMap    map[string]int
-		addKeys      map[string]int
-		deleteKeys   []string
-		wantModified map[string]int
-		wantDeleted  map[string]bool
-	}{
-		"SourceMapOnlyNoModifications": {
-			sourceMap:    map[string]int{"k1": 1, "k2": 2},
-			addKeys:      map[string]int{},
-			deleteKeys:   []string{},
-			wantModified: map[string]int{"k1": 1, "k2": 2},
-			wantDeleted:  map[string]bool{},
-		},
-		"NilSourceMapAddAndDelete": {
-			sourceMap:    nil,
-			addKeys:      map[string]int{"a": 1, "b": 2},
-			deleteKeys:   []string{"b"},
-			wantModified: map[string]int{"a": 1},
-			wantDeleted:  map[string]bool{"b": true},
-		},
-		"EmptySourceMapAddAndDelete": {
-			sourceMap:    map[string]int{},
-			addKeys:      map[string]int{"x": 10},
-			deleteKeys:   []string{"y"},
-			wantModified: map[string]int{"x": 10},
-			wantDeleted:  map[string]bool{"y": true},
-		},
-		"NonEmptySourceMapAddOverwriteDelete": {
-			sourceMap:    map[string]int{"orig1": 100, "orig2": 200},
-			addKeys:      map[string]int{"new1": 300, "orig1": 101},
-			deleteKeys:   []string{"orig2", "new1"},
-			wantModified: map[string]int{"orig1": 101},
-			wantDeleted:  map[string]bool{"orig2": true, "new1": true},
-		},
-		"DeleteKeyFromSourceMap": {
-			sourceMap:    map[string]int{"key_to_delete": 70, "key_to_keep": 80},
-			addKeys:      map[string]int{},
-			deleteKeys:   []string{"key_to_delete"},
-			wantModified: map[string]int{"key_to_keep": 80},
-			wantDeleted:  map[string]bool{"key_to_delete": true},
-		},
-		"AddOnlyNoSourceMap": {
-			sourceMap:    nil,
-			addKeys:      map[string]int{"add1": 10, "add2": 20},
-			deleteKeys:   []string{},
-			wantModified: map[string]int{"add1": 10, "add2": 20},
-			wantDeleted:  map[string]bool{},
-		},
-		"DeleteOnlyNoSourceMap": {
-			sourceMap:    nil,
-			addKeys:      map[string]int{},
-			deleteKeys:   []string{"del1", "del2"},
-			wantModified: map[string]int{},
-			wantDeleted:  map[string]bool{"del1": true, "del2": true},
-		},
-		"DeleteKeyNotPresentInSourceOrAdded": {
-			sourceMap:    map[string]int{"a": 1},
-			addKeys:      map[string]int{"b": 2},
-			deleteKeys:   []string{"c"},
-			wantModified: map[string]int{"a": 1, "b": 2},
-			wantDeleted:  map[string]bool{"c": true},
-		},
-		"AddKeyThenDeleteIt": {
-			sourceMap:    map[string]int{"base": 100},
-			addKeys:      map[string]int{"temp": 50},
-			deleteKeys:   []string{"temp"},
-			wantModified: map[string]int{"base": 100},
-			wantDeleted:  map[string]bool{"temp": true},
-		},
-		"AllOperations": {
-			sourceMap:    map[string]int{"s1": 1, "s2": 2, "s3": 3},
-			addKeys:      map[string]int{"a1": 10, "s2": 22, "a2": 20},
-			deleteKeys:   []string{"s1", "a1", "nonexistent"},
-			wantModified: map[string]int{"s2": 22, "s3": 3, "a2": 20},
-			wantDeleted:  map[string]bool{"s1": true, "a1": true, "nonexistent": true},
-		},
-		"NoOperationsEmptySource": {
-			sourceMap:    map[string]int{},
-			addKeys:      map[string]int{},
-			deleteKeys:   []string{},
-			wantModified: map[string]int{},
-			wantDeleted:  map[string]bool{},
-		},
-	}
-
-	for testName, test := range tests {
-		t.Run(testName, func(t *testing.T) {
-			p := newPatchFromMap(test.sourceMap)
-
-			for k, v := range test.addKeys {
-				p.Set(k, v)
-			}
-			for _, k := range test.deleteKeys {
-				p.Delete(k)
-			}
-
-			if !maps.Equal(p.Modified, test.wantModified) {
-				t.Errorf("Modified map mismatch: got %v, want %v", p.Modified, test.wantModified)
-			}
-
-			if !maps.Equal(p.Deleted, test.wantDeleted) {
-				t.Errorf("Deleted map mismatch: got %v, want %v", p.Deleted, test.wantDeleted)
-			}
-		})
-	}
-}
-
-func TestMergePatchesInPlace(t *testing.T) {
-	tests := map[string]struct {
-		modifiedPatchA     map[string]int
-		deletedPatchA      map[string]bool
-		modifiedPatchB     map[string]int
-		deletedPatchB      map[string]bool
-		wantModifiedMerged map[string]int
-		wantDeletedMerged  map[string]bool
-	}{
-		"PatchBOverwritesValueInPatchA": {
-			modifiedPatchA:     map[string]int{"a": 1, "b": 2},
-			deletedPatchA:      map[string]bool{},
-			modifiedPatchB:     map[string]int{"b": 22, "c": 3},
-			deletedPatchB:      map[string]bool{},
-			wantModifiedMerged: map[string]int{"a": 1, "b": 22, "c": 3},
-			wantDeletedMerged:  map[string]bool{},
-		},
-		"PatchBDeletesValuePresentInPatchA": {
-			modifiedPatchA:     map[string]int{"a": 1, "b": 2},
-			deletedPatchA:      map[string]bool{},
-			modifiedPatchB:     map[string]int{},
-			deletedPatchB:      map[string]bool{"a": true},
-			wantModifiedMerged: map[string]int{"b": 2},
-			wantDeletedMerged:  map[string]bool{"a": true},
-		},
-		"PatchBDeletesValueAlreadyDeletedInPatchA": {
-			modifiedPatchA:     map[string]int{},
-			deletedPatchA:      map[string]bool{"x": true},
-			modifiedPatchB:     map[string]int{},
-			deletedPatchB:      map[string]bool{"x": true, "y": true},
-			wantModifiedMerged: map[string]int{},
-			wantDeletedMerged:  map[string]bool{"x": true, "y": true},
-		},
-		"PatchBAddsValuePreviouslyDeletedInPatchA": {
-			modifiedPatchA:     map[string]int{},
-			deletedPatchA:      map[string]bool{"x": true},
-			modifiedPatchB:     map[string]int{"x": 10},
-			deletedPatchB:      map[string]bool{},
-			wantModifiedMerged: map[string]int{"x": 10},
-			wantDeletedMerged:  map[string]bool{},
-		},
-		"MergeEmptyPatchBIntoNonEmptyPatchA": {
-			modifiedPatchA:     map[string]int{"a": 1},
-			deletedPatchA:      map[string]bool{"b": true},
-			modifiedPatchB:     map[string]int{},
-			deletedPatchB:      map[string]bool{},
-			wantModifiedMerged: map[string]int{"a": 1},
-			wantDeletedMerged:  map[string]bool{"b": true},
-		},
-		"MergeNonEmptyPatchBIntoEmptyPatchA": {
-			modifiedPatchA:     map[string]int{},
-			deletedPatchA:      map[string]bool{},
-			modifiedPatchB:     map[string]int{"a": 1},
-			deletedPatchB:      map[string]bool{"b": true},
-			wantModifiedMerged: map[string]int{"a": 1},
-			wantDeletedMerged:  map[string]bool{"b": true},
-		},
-		"MergeTwoEmptyPatches": {
-			modifiedPatchA:     map[string]int{},
-			deletedPatchA:      map[string]bool{},
-			modifiedPatchB:     map[string]int{},
-			deletedPatchB:      map[string]bool{},
-			wantModifiedMerged: map[string]int{},
-			wantDeletedMerged:  map[string]bool{},
-		},
-		"NoOverlapBetweenPatchAAndPatchBModifications": {
-			modifiedPatchA:     map[string]int{"a1": 1, "a2": 2},
-			deletedPatchA:      map[string]bool{},
-			modifiedPatchB:     map[string]int{"b1": 10, "b2": 20},
-			deletedPatchB:      map[string]bool{},
-			wantModifiedMerged: map[string]int{"a1": 1, "a2": 2, "b1": 10, "b2": 20},
-			wantDeletedMerged:  map[string]bool{},
-		},
-		"NoOverlapBetweenPatchAAndPatchBDeletions": {
-			modifiedPatchA:     map[string]int{},
-			deletedPatchA:      map[string]bool{"a1": true, "a2": true},
-			modifiedPatchB:     map[string]int{},
-			deletedPatchB:      map[string]bool{"b1": true, "b2": true},
-			wantModifiedMerged: map[string]int{},
-			wantDeletedMerged:  map[string]bool{"a1": true, "a2": true, "b1": true, "b2": true},
-		},
-		"PatchBOnlyAddsNewKeysPatchAUnchanged": {
-			modifiedPatchA:     map[string]int{"orig": 5},
-			modifiedPatchB:     map[string]int{"new1": 100, "new2": 200},
-			deletedPatchA:      map[string]bool{},
-			deletedPatchB:      map[string]bool{},
-			wantModifiedMerged: map[string]int{"orig": 5, "new1": 100, "new2": 200},
-			wantDeletedMerged:  map[string]bool{},
-		},
-		"PatchBOnlyDeletesNewKeysPatchAUnchanged": {
-			modifiedPatchA:     map[string]int{"orig": 5},
-			deletedPatchA:      map[string]bool{},
-			modifiedPatchB:     map[string]int{},
-			deletedPatchB:      map[string]bool{"del1": true, "del2": true},
-			wantModifiedMerged: map[string]int{"orig": 5},
-			wantDeletedMerged:  map[string]bool{"del1": true, "del2": true},
-		},
-		"AllInOne": {
-			modifiedPatchA:     map[string]int{"k1": 1, "k2": 2, "k3": 3},
-			deletedPatchA:      map[string]bool{"d1": true, "d2": true},
-			modifiedPatchB:     map[string]int{"k2": 22, "k4": 4, "d1": 11},
-			deletedPatchB:      map[string]bool{"k3": true, "d2": true, "d3": true},
-			wantModifiedMerged: map[string]int{"k1": 1, "k2": 22, "k4": 4, "d1": 11},
-			wantDeletedMerged:  map[string]bool{"k3": true, "d2": true, "d3": true},
-		},
-	}
-
-	for testName, test := range tests {
-		t.Run(testName, func(t *testing.T) {
-			patchA := newPatchFromMap(test.modifiedPatchA)
-			for k := range test.deletedPatchA {
-				patchA.Delete(k)
-			}
-
-			patchB := newPatchFromMap(test.modifiedPatchB)
-			for k := range test.deletedPatchB {
-				patchB.Delete(k)
-			}
-
-			merged := mergePatchesInPlace(patchA, patchB)
-
-			if merged != patchA {
-				t.Errorf("mergePatchesInPlace did not modify patchA inplace, references are different")
-			}
-
-			if !maps.Equal(merged.Modified, test.wantModifiedMerged) {
-				t.Errorf("Modified map mismatch: got %v, want %v", merged.Modified, test.wantModifiedMerged)
-			}
-
-			if !maps.Equal(merged.Deleted, test.wantDeletedMerged) {
-				t.Errorf("Deleted map mismatch: got %v, want %v", merged.Deleted, test.wantDeletedMerged)
-			}
-		})
-	}
-}
 
 func TestPatchSetAsMap(t *testing.T) {
 	tests := map[string]struct {
@@ -572,7 +270,7 @@ func TestPatchSetRevert(t *testing.T) {
 
 func TestPatchSetForkRevert(t *testing.T) {
 	// 1. Initialize empty patchset
-	ps := newPatchSet[string, int](newPatch[string, int]())
+	ps := NewPatchSet(NewPatch[string, int]())
 	if initialMap := ps.AsMap(); len(initialMap) != 0 {
 		t.Fatalf("Initial AsMap() got %v, want empty", initialMap)
 	}
@@ -602,7 +300,7 @@ func TestPatchSetForkRevert(t *testing.T) {
 
 func TestPatchSetForkCommit(t *testing.T) {
 	// 1. Initialize empty patchset
-	ps := newPatchSet[string, int](newPatch[string, int]())
+	ps := NewPatchSet(NewPatch[string, int]())
 	if initialMap := ps.AsMap(); len(initialMap) != 0 {
 		t.Fatalf("Initial AsMap() got %v, want empty", initialMap)
 	}
@@ -726,13 +424,13 @@ func TestPatchSetFindValue(t *testing.T) {
 func TestPatchSetOperations(t *testing.T) {
 	tests := map[string]struct {
 		patchLayers    []map[string]*int
-		mutatePatchSet func(ps *patchSet[string, int])
+		mutatePatchSet func(ps *PatchSet[string, int])
 		searchKey      string
 		wantInCurrent  bool
 	}{
 		"SetCurrent_NewKey_InitiallyEmptyPatchSet": {
 			patchLayers: []map[string]*int{},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.SetCurrent("a", 1)
 			},
 			searchKey:     "a",
@@ -740,7 +438,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"SetCurrent_NewKey_OnExistingEmptyLayer": {
 			patchLayers: []map[string]*int{{}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.SetCurrent("a", 1)
 			},
 			searchKey:     "a",
@@ -748,7 +446,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"SetCurrent_OverwriteKey_InCurrentPatch": {
 			patchLayers: []map[string]*int{{"a": ptr.To(10)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.SetCurrent("a", 1)
 			},
 			searchKey:     "a",
@@ -756,7 +454,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"SetCurrent_OverwriteKey_InLowerPatch_AfterFork": {
 			patchLayers: []map[string]*int{{"a": ptr.To(10)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.Fork()
 				ps.SetCurrent("a", 1)
 			},
@@ -765,7 +463,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"DeleteCurrent_ExistingKey_InCurrentPatch": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.DeleteCurrent("a")
 			},
 			searchKey:     "a",
@@ -773,7 +471,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"DeleteCurrent_ExistingKey_InLowerPatch_AfterFork": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.Fork()
 				ps.DeleteCurrent("a")
 			},
@@ -782,7 +480,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"DeleteCurrent_NonExistentKey_OnExistingEmptyLayer": {
 			patchLayers: []map[string]*int{{}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.DeleteCurrent("x")
 			},
 			searchKey:     "x",
@@ -790,7 +488,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"DeleteCurrent_NonExistentKey_InitiallyEmptyPatchSet": {
 			patchLayers: []map[string]*int{}, // Starts with no layers
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.DeleteCurrent("x")
 			},
 			searchKey:     "x",
@@ -798,19 +496,19 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"InCurrentPatch_KeySetInCurrent": {
 			patchLayers:    []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {},
+			mutatePatchSet: func(ps *PatchSet[string, int]) {},
 			searchKey:      "a",
 			wantInCurrent:  true,
 		},
 		"InCurrentPatch_KeySetInLower_NotInCurrent_AfterFork": {
 			patchLayers:    []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) { ps.Fork() },
+			mutatePatchSet: func(ps *PatchSet[string, int]) { ps.Fork() },
 			searchKey:      "a",
 			wantInCurrent:  false,
 		},
 		"InCurrentPatch_KeyDeletedInCurrent": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.DeleteCurrent("a")
 			},
 			searchKey:     "a",
@@ -818,7 +516,7 @@ func TestPatchSetOperations(t *testing.T) {
 		},
 		"InCurrentPatch_NonExistentKey_InitiallyEmptyPatchSet": {
 			patchLayers:    []map[string]*int{},
-			mutatePatchSet: func(ps *patchSet[string, int]) {},
+			mutatePatchSet: func(ps *PatchSet[string, int]) {},
 			searchKey:      "a",
 			wantInCurrent:  false,
 		},
@@ -840,25 +538,25 @@ func TestPatchSetOperations(t *testing.T) {
 func TestPatchSetCache(t *testing.T) {
 	tests := map[string]struct {
 		patchLayers     []map[string]*int
-		mutatePatchSet  func(ps *patchSet[string, int])
+		mutatePatchSet  func(ps *PatchSet[string, int])
 		wantCache       map[string]*int
 		wantCacheInSync bool
 	}{
 		"Initial_EmptyPatchSet": {
 			patchLayers:     []map[string]*int{},
-			mutatePatchSet:  func(ps *patchSet[string, int]) {},
+			mutatePatchSet:  func(ps *PatchSet[string, int]) {},
 			wantCache:       map[string]*int{},
 			wantCacheInSync: false,
 		},
 		"Initial_WithData_NoCacheAccess": {
 			patchLayers:     []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet:  func(ps *patchSet[string, int]) {},
+			mutatePatchSet:  func(ps *PatchSet[string, int]) {},
 			wantCache:       map[string]*int{},
 			wantCacheInSync: false,
 		},
 		"FindValue_PopulatesCacheForKey": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1), "b": ptr.To(2)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.FindValue("a")
 			},
 			wantCache:       map[string]*int{"a": ptr.To(1)},
@@ -866,7 +564,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"FindValue_DeletedKey_PopulatesCacheWithNil": {
 			patchLayers: []map[string]*int{{"a": nil, "b": ptr.To(2)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.FindValue("a")
 			},
 			wantCache:       map[string]*int{"a": nil},
@@ -874,7 +572,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"AsMap_PopulatesAndSyncsCache": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1), "b": nil, "c": ptr.To(3)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.AsMap()
 			},
 			wantCache:       map[string]*int{"a": ptr.To(1), "c": ptr.To(3)}, // Cache does not necessarily track deletions like 'b' key
@@ -882,7 +580,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"SetCurrent_UpdatesCache_NewKey": {
 			patchLayers: []map[string]*int{{}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.SetCurrent("x", 10)
 			},
 			wantCache:       map[string]*int{"x": ptr.To(10)},
@@ -890,7 +588,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"SetCurrent_UpdatesCache_OverwriteKey": {
 			patchLayers: []map[string]*int{{"x": ptr.To(5)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.FindValue("x")
 				ps.SetCurrent("x", 10)
 			},
@@ -899,7 +597,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"DeleteCurrent_UpdatesCache": {
 			patchLayers: []map[string]*int{{"x": ptr.To(5)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.FindValue("x")
 				ps.DeleteCurrent("x")
 			},
@@ -908,7 +606,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"Revert_ClearsAffectedCacheEntries_And_SetsCacheNotInSync": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}, {"b": ptr.To(2), "a": ptr.To(11)}}, // Layer 0: a=1; Layer 1: b=2, a=11
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.FindValue("a")
 				ps.FindValue("b")
 				ps.Revert()
@@ -918,7 +616,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"Revert_OnSyncedCache_SetsCacheNotInSync": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}, {"b": ptr.To(2)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.AsMap()
 				ps.Revert()
 			},
@@ -927,7 +625,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"Commit_DoesNotInvalidateCache_IfValuesConsistent": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}, {"b": ptr.To(2)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.FindValue("a")
 				ps.FindValue("b")
 				ps.Commit()
@@ -937,7 +635,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"Commit_OnSyncedCache_KeepsCacheInSync": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}, {"b": ptr.To(2)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.AsMap()
 				ps.Commit()
 			},
@@ -946,7 +644,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"Fork_DoesNotInvalidateCache": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.FindValue("a")
 				ps.Fork()
 			},
@@ -955,7 +653,7 @@ func TestPatchSetCache(t *testing.T) {
 		},
 		"Fork_OnSyncedCache_KeepsCacheInSync": {
 			patchLayers: []map[string]*int{{"a": ptr.To(1)}},
-			mutatePatchSet: func(ps *patchSet[string, int]) {
+			mutatePatchSet: func(ps *PatchSet[string, int]) {
 				ps.AsMap()
 				ps.Fork()
 			},
@@ -988,14 +686,14 @@ func TestPatchSetCache(t *testing.T) {
 	}
 }
 
-func buildTestPatchSet[K comparable, V any](t *testing.T, patchLayers []map[K]*V) *patchSet[K, V] {
+func buildTestPatchSet[K comparable, V any](t *testing.T, patchLayers []map[K]*V) *PatchSet[K, V] {
 	t.Helper()
 
 	patchesNumber := len(patchLayers)
-	patches := make([]*patch[K, V], patchesNumber)
+	patches := make([]*Patch[K, V], patchesNumber)
 	for i := 0; i < patchesNumber; i++ {
 		layerMap := patchLayers[i]
-		currentPatch := newPatch[K, V]()
+		currentPatch := NewPatch[K, V]()
 		for k, vPtr := range layerMap {
 			if vPtr != nil {
 				currentPatch.Set(k, *vPtr)
@@ -1006,5 +704,5 @@ func buildTestPatchSet[K comparable, V any](t *testing.T, patchLayers []map[K]*V
 		patches[i] = currentPatch
 	}
 
-	return newPatchSet(patches...)
+	return NewPatchSet(patches...)
 }
